requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
psycopg2==2.8.4
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.7
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
